Arrange things more nicely.
