+++ tests/commands.go
@@ -3,35 +3,9 @@ package tests
 import (
 	"fmt"
 	"github.com/tmr232/goat"
-	"io"
-	"reflect"
 )
 
 //go:generate go run github.com/tmr232/goat/cmd/goater
-
-var testCommandWriter io.Writer
-
-type writerContext struct {
-	io.Writer
-}
-
-func (c writerContext) restore() {
-	testCommandWriter = c
-}
-
-func withWriter(writer io.Writer) writerContext {
-	original := testCommandWriter
-	testCommandWriter = writer
-	return writerContext{original}
-}
-
-func printArg(name string, value any) {
-	if reflect.TypeOf(value).Kind() != reflect.Pointer || reflect.ValueOf(value).IsNil() {
-		fmt.Fprintf(testCommandWriter, "%s = %#v\n", name, value)
-	} else {
-		fmt.Fprintf(testCommandWriter, "%s = *-> %#v\n", name, reflect.ValueOf(value).Elem().Interface())
-	}
-}
 
 func withIntFlags(required, defaultValue int, optional *int) {
 	goat.Flag(defaultValue).Default(42)
